05/15 - added logic for PUT and POST capacity check - cleaned up code and added throws to service class methods

+++ src/Project 2/warehouse-inventory-api/src/main/java/com/skillstorm/warehouseinventoryapi/Services/BranchService.java
@@ -4,50 +4,63 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.skillstorm.warehouseinventoryapi.models.Branch;
+
 import com.skillstorm.warehouseinventoryapi.repos.BranchRepo;
 
 
- // make transactional?
 
 
-
+@Transactional // spring framework annotation or jakarta?
 @Service
 public class BranchService {
 
-      private BranchRepo repo;
+    private BranchRepo repo;
 
-    // @Autowired // ?
+
     public BranchService(BranchRepo repo) {
         this.repo = repo;
     }
 
-    
+    // return all branches
     public List<Branch> findAll() {
-        return repo.findAll(); // return all branches
+        return repo.findAll();
     }
 
+    // return branch name and number
     public List<Branch> findByName(String name) {
+
         return repo.findByName(name); // return branch name and number
     }
 
-    public Optional<Branch> findbyId(int id) { // added optional util
+    // returns Branch data by ID
+    public Optional<Branch> findbyId(int id) {
+        if (!repo.existsById(id))
+            throw new IllegalArgumentException("No Branch with id " + id + " found"); // added optional util
         return repo.findById(id); // returns Branch data by ID
     }
 
-    public Branch create(Branch newBranch) { // needs authorization permissions - super user
-        return newBranch; // returns newly created branch (name and ID)
+    // returns newly created branch
+    public Branch create(Branch newBranch) { // add authorization permissions - super user
+        return repo.save(newBranch);
     }
 
-  public Branch update(int id, Branch product) {
-        Optional<Branch> currentItem = repo.findById(id); // add logic to update fields
-        return repo.save(currentItem); // returns updated branch
+    // returns updated branch
+    public Branch update(int id, Branch branch) { // add authorization permissions - super user
+        if (!repo.existsById(id))
+            throw new IllegalArgumentException("No Branch with id " + id + "exists");
+        Optional<Branch> updatedBranch = repo.findById(id); // add logic to update fields
+        return repo.save(updatedBranch);
+
     }
-    public void deleteById(int id) { // removes branch by ID
-        // check if ID exists and throw exception
-        // ideally perhaps return a list of current branches to confirm
+
+    // Deletes branch by ID
+    public void deleteById(int id) { // add authorization permissions - super user
+        if (!repo.existsById(id))
+            throw new IllegalArgumentException("No Branch with id " + id + "exists");
         repo.deleteById(id);
     }
-   
+
 }
